+++ src/internal_events/open.rs
@@ -42,6 +42,7 @@ impl OpenGauge {
         }
     }
 
+    #[cfg(feature = "sources-utils-unix")]
     pub fn any_open(&self) -> bool {
         self.gauge.load(Ordering::Acquire) != 0
     }
